fixed getObjectPerceptsForAgent
Okay, so the issue was that (1) we didn't provide the coordinates relative to the agent with respect to rotation and position, and (2) sometimes the FieldOfView isInRange implementations calculation different at the 10+th decimal place and therefore disagreed with our collision calculations. I fixed this by simply adding it as a second filter; might not be pretty and efficient but it works.