Add: dynamic priority list (does not fix stuck issue!)
Once we enter a door, the priority is indeed lowered putting Door at the
bottom of the priority list. But even then, the agent is stuck. After
some debugging I'm thinking it might have to do with the fact that the
range is decreased and that spells trouble for isInsideOtherVertex.
MAYBE. I'm not sure. For today, I give up on debugging it more. Perhaps
it's not worth it if we plan on rewritting anyway.